Update API to directly return feature collection. Update schema to match OGC spec.

+++ SOH.Process.Server/Simulations/Jobs/SimulationRunJobHandler.cs
@@ -1,5 +1,9 @@
+using Mars.Common;
+using Mars.Common.Core.Collections;
 using Mars.Components.Starter;
+using Mars.Core.Data;
 using Mars.Core.Simulation;
+using Mars.Core.Simulation.Entities;
 using Mars.Interfaces.Model;
 using Mars.Interfaces.Model.Options;
 using MediatR;
@@ -10,6 +14,7 @@ using SOH.Process.Server.Models.Common.Exceptions;
 using SOH.Process.Server.Models.Ogc;
 using SOH.Process.Server.Models.Processes;
 using SOH.Process.Server.Resources;
+using SOH.Process.Server.Simulations.Workflows;
 using SOHModel.Domain.Graph;
 using SOHModel.Ferry.Model;
 using SOHModel.Ferry.Route;
@@ -21,6 +26,7 @@ namespace SOH.Process.Server.Simulations.Jobs;
 public class SimulationRunJobHandler(
     ISimulationService simulationService,
     IResultService resultService,
+    JobSubscribeWorkflow jobSubscribeWorkflow,
     IStringLocalizer<SharedResource> localization)
     : IRequestHandler<SimulationRunJobRequest, Unit>
 {
@@ -28,7 +34,7 @@ public class SimulationRunJobHandler(
     {
         var job = await simulationService.FindJobAsync(request.JobId, cancellationToken);
         var process = await simulationService.FindSimulationAsync(
-            job?.SimulationId ?? string.Empty, cancellationToken);
+            job?.ProcessId ?? string.Empty, cancellationToken);
 
         if (process == null || job == null) return Unit.Value;
 
@@ -49,12 +55,11 @@ public class SimulationRunJobHandler(
         var currentJob = await simulationService.GetSimulationJobAsync(request.JobId, cancellationToken);
         currentJob.FinishedUtc = DateTime.UtcNow;
         await simulationService.UpdateAsync(currentJob.JobId, currentJob, cancellationToken);
-
         return Unit.Value;
     }
 
     private async Task ExecuteTestProcess(SimulationRunJobRequest request,
-        string jobId, SimulationProcess process, CancellationToken cancellationToken)
+        string jobId, SimulationProcessDescription processDescription, CancellationToken cancellationToken)
     {
         var currentJob = await simulationService.GetSimulationJobAsync(jobId, cancellationToken);
         try
@@ -63,9 +68,9 @@ public class SimulationRunJobHandler(
             {
 #if DEBUG
                 if (request.Execute != null && request.Execute.Inputs.TryGetValue("func",
-                        out object? function) && function is Action action)
+                        out object? function) && function is Action<int, SimulationJob> action)
                 {
-                    action.Invoke();
+                    action.Invoke(i, currentJob);
                 }
 #endif
                 await Task.Delay(100, cancellationToken);
@@ -76,7 +81,7 @@ public class SimulationRunJobHandler(
             currentJob.Status = StatusCode.Successful;
             currentJob.ResultId = await resultService.CreateAsync(new Result
             {
-                SimulationId = process.Id,
+                ProcessId = processDescription.Id,
                 JobId = currentJob.JobId,
                 FeatureCollection =
                 [
@@ -91,12 +96,13 @@ public class SimulationRunJobHandler(
         {
             currentJob.Status = StatusCode.Failed;
         }
-
         await simulationService.UpdateAsync(currentJob.JobId, currentJob, cancellationToken);
+
+        await jobSubscribeWorkflow.NotifySubscriberAsync(request, currentJob.JobId, cancellationToken);
     }
 
     private async Task ExecuteSimulationProcess(ModelDescription description, SimulationConfig simConfig,
-        string jobId, SimulationProcess process, CancellationToken token)
+        string jobId, SimulationProcessDescription processDescription, CancellationToken token)
     {
         var currentJob = await simulationService.GetSimulationJobAsync(jobId, token);
         try
@@ -106,8 +112,13 @@ public class SimulationRunJobHandler(
             var simulation = application.Resolve<ISimulation>();
             var step = simulation.PrepareSimulation(description, simConfig);
 
+            var serializer = application.Resolve<ISerializerManager>();
+
+            var featureCollection = new FeatureCollection();
             while (!step.IsFinished)
             {
+                CollectResultForFeatureCollection(serializer, step, featureCollection);
+
                 if (currentJob.IsCancellationRequested)
                 {
                     break;
@@ -121,41 +132,72 @@ public class SimulationRunJobHandler(
                 currentJob = await simulationService.GetSimulationJobAsync(jobId, token);
             }
 
-            string[] geoJsonFiles = Directory.GetFiles(
-                $"results_{simConfig.SimulationIdentifier}", "*.geojson");
-
-            var featureCollection = new FeatureCollection();
-            var geoJsonReader = new NetTopologySuite.IO.GeoJsonReader();
-            foreach (string geoJsonFile in geoJsonFiles.Where(path => new FileInfo(path).Length > 0))
-            {
-                // var singleCollection = geoJsonReader.Read<FeatureCollection>(geoJsonFile);
-                //
-                // foreach (var feature in singleCollection)
-                // {
-                //     featureCollection.Add(feature);
-                // }
-            }
+            // string[] geoJsonFiles = Directory.GetFiles(
+            // $"results_{simConfig.SimulationIdentifier}", "*.geojson");
 
             var result = new Result
             {
-                SimulationId = process.Id, JobId = currentJob.JobId,
+                ProcessId = processDescription.Id, JobId = currentJob.JobId,
                 FeatureCollection = featureCollection
             };
 
             currentJob.ResultId = await resultService.CreateAsync(result, token);
             currentJob.Status = currentJob.IsCancellationRequested ? StatusCode.Dismissed : StatusCode.Successful;
         }
-        catch
+        catch(Exception exception)
         {
+            currentJob.ExceptionMessage = exception.Message;
             currentJob.Status = StatusCode.Failed;
         }
 
         await simulationService.UpdateAsync(currentJob.JobId, currentJob, token);
-        Directory.Delete($"results_{simConfig.SimulationIdentifier}", true);
+
+        if (Directory.Exists($"results_{simConfig.SimulationIdentifier}"))
+        {
+            Directory.Delete($"results_{simConfig.SimulationIdentifier}", true);
+        }
+
+    }
+
+    private static void CollectResultForFeatureCollection(ISerializerManager serializer, SimulationWorkflowState step,
+        FeatureCollection featureCollection)
+    {
+        var typeLoggers = serializer
+            .GetTypeLoggers()
+            .Where(logger => logger.Mapping is AgentMapping);
+
+        foreach (var typeLogger in typeLoggers)
+        {
+            long currentTick = step.CurrentTick;
+            var serializeAbleProxies =
+                typeLogger.EntityProxies.Where(proxy =>
+                    (proxy.OutputTicks != null && proxy.OutputTicks.Contains(currentTick)) ||
+                    (proxy.OutputFrequency > 0 && currentTick % proxy.OutputFrequency == 0));
+
+            foreach (var entityLogger in serializeAbleProxies.Where(logger => logger.Position != null))
+            {
+                var point = new Point(entityLogger.Position.X, entityLogger.Position.Y);
+                var dictionary = entityLogger.SerializeProperties()
+                    .ToDictionary(tuple => tuple.Item1, tuple => tuple.Item2);
+                var feature = new Feature(point, new AttributesTable(dictionary));
+
+                featureCollection.Add(feature);
+
+                if (featureCollection.BoundingBox == null)
+                {
+                    featureCollection.BoundingBox = feature.BoundingBox;
+                }
+                else
+                {
+                    featureCollection.BoundingBox.ExpandToInclude(point.X, point.Y);
+                }
+            }
+        }
     }
 
     private async Task<(ModelDescription description, SimulationConfig simConfig)> ConstructModelAndConfig(
-        SimulationRunJobRequest request, SimulationProcess process, SimulationJob job, CancellationToken cancellationToken)
+        SimulationRunJobRequest request, SimulationProcessDescription processDescription,
+        SimulationJob job, CancellationToken cancellationToken)
     {
         var description = new ModelDescription();
         description.AddLayer<FerryLayer>();
@@ -170,12 +212,12 @@ public class SimulationRunJobHandler(
         description.AddAgent<DockWorker, DockWorkerLayer>();
         description.AddEntity<Ferry>();
 
-        var simConfig = await GetSimulationConfigOrDefault(process, job, request.Execute, cancellationToken);
+        var simConfig = await GetSimulationConfigOrDefault(processDescription, job, request.Execute, cancellationToken);
         return (description, simConfig);
     }
 
     private async Task<SimulationConfig> GetSimulationConfigOrDefault(
-        SimulationProcess process, SimulationJob job,
+        SimulationProcessDescription processDescription, SimulationJob job,
         Execute? processConfig, CancellationToken token)
     {
         SimulationConfig config;
@@ -190,7 +232,7 @@ public class SimulationRunJobHandler(
             {
                 await simulationService.DeleteAsync(job.JobId, token);
                 throw new BadRequestException(
-                    localization[$"Invalid 'config' input for selected process '{process.Title}'"], [ex.Message]);
+                    localization[$"Invalid 'config' input for selected process '{processDescription.Title}'"], [ex.Message]);
             }
         }
         else
@@ -211,7 +253,14 @@ public class SimulationRunJobHandler(
         {
             mapping.LiveVisualization = null;
 
-            mapping.OutputTarget = mapping is AgentMapping ? OutputTargetType.GeoJsonFile : OutputTargetType.None;
+            if (mapping is AgentMapping)
+            {
+                mapping.OutputTarget = OutputTargetType.Manual;
+            }
+            else
+            {
+                mapping.OutputTarget = OutputTargetType.None;
+            }
         }
 
         return config;
